third_party/blink/web_tests/platform/mac-mac11/external/wpt/css/css-values/clamp-length-serialize-expected.txt

This is a testharness.js-based test.
FAIL e.style['letter-spacing'] = "clamp(1px, 2px, 3px)" should set the property value assert_equals: serialization should be canonical expected "calc(2px)" but got "clamp(1px, 2px, 3px)"
FAIL e.style['letter-spacing'] = "calc(clamp(1px, 2px, 3px))" should set the property value assert_equals: serialization should be canonical expected "calc(2px)" but got "clamp(1px, 2px, 3px)"
FAIL e.style['letter-spacing'] = "clamp(1px, 2px, clamp(2px, 3px, 4px))" should set the property value assert_equals: serialization should be canonical expected "calc(2px)" but got "clamp(1px, 2px, clamp(2px, 3px, 4px))"
FAIL e.style['letter-spacing'] = "calc(clamp(1px, 2px, clamp(2px, 3px, 4px)))" should set the property value assert_equals: serialization should be canonical expected "calc(2px)" but got "clamp(1px, 2px, clamp(2px, 3px, 4px))"
FAIL e.style['letter-spacing'] = "clamp(10px, 5px, 30px)" should set the property value assert_equals: serialization should be canonical expected "calc(10px)" but got "clamp(10px, 5px, 30px)"
FAIL e.style['letter-spacing'] = "calc(clamp(10px, 5px, 30px))" should set the property value assert_equals: serialization should be canonical expected "calc(10px)" but got "clamp(10px, 5px, 30px)"
FAIL e.style['letter-spacing'] = "clamp(10px, 35px, 30px)" should set the property value assert_equals: serialization should be canonical expected "calc(30px)" but got "clamp(10px, 35px, 30px)"
FAIL e.style['letter-spacing'] = "calc(clamp(10px, 35px, 30px))" should set the property value assert_equals: serialization should be canonical expected "calc(30px)" but got "clamp(10px, 35px, 30px)"
FAIL e.style['letter-spacing'] = "clamp(10px, 35px , 30px)" should set the property value assert_equals: serialization should be canonical expected "calc(30px)" but got "clamp(10px, 35px, 30px)"
FAIL e.style['letter-spacing'] = "calc(clamp(10px, 35px , 30px))" should set the property value assert_equals: serialization should be canonical expected "calc(30px)" but got "clamp(10px, 35px, 30px)"
FAIL e.style['letter-spacing'] = "clamp(10px, 35px /*foo*/, 30px)" should set the property value assert_equals: serialization should be canonical expected "calc(30px)" but got "clamp(10px, 35px, 30px)"
FAIL e.style['letter-spacing'] = "calc(clamp(10px, 35px /*foo*/, 30px))" should set the property value assert_equals: serialization should be canonical expected "calc(30px)" but got "clamp(10px, 35px, 30px)"
FAIL e.style['letter-spacing'] = "clamp(10px /* foo */ , 35px, 30px)" should set the property value assert_equals: serialization should be canonical expected "calc(30px)" but got "clamp(10px, 35px, 30px)"
FAIL e.style['letter-spacing'] = "calc(clamp(10px /* foo */ , 35px, 30px))" should set the property value assert_equals: serialization should be canonical expected "calc(30px)" but got "clamp(10px, 35px, 30px)"
FAIL e.style['letter-spacing'] = "clamp(10px , 35px, 30px)" should set the property value assert_equals: serialization should be canonical expected "calc(30px)" but got "clamp(10px, 35px, 30px)"
FAIL e.style['letter-spacing'] = "calc(clamp(10px , 35px, 30px))" should set the property value assert_equals: serialization should be canonical expected "calc(30px)" but got "clamp(10px, 35px, 30px)"
FAIL e.style['letter-spacing'] = "clamp(30px, 100px, 20px)" should set the property value assert_equals: serialization should be canonical expected "calc(30px)" but got "clamp(30px, 100px, 20px)"
FAIL e.style['letter-spacing'] = "calc(clamp(30px, 100px, 20px))" should set the property value assert_equals: serialization should be canonical expected "calc(30px)" but got "clamp(30px, 100px, 20px)"
FAIL e.style['letter-spacing'] = "clamp(-30px, -20px, -10px)" should set the property value assert_equals: serialization should be canonical expected "calc(-20px)" but got "clamp(-30px, -20px, -10px)"
FAIL e.style['letter-spacing'] = "calc(clamp(-30px, -20px, -10px))" should set the property value assert_equals: serialization should be canonical expected "calc(-20px)" but got "clamp(-30px, -20px, -10px)"
FAIL e.style['letter-spacing'] = "clamp(-30px, -100px, -10px)" should set the property value assert_equals: serialization should be canonical expected "calc(-30px)" but got "clamp(-30px, -100px, -10px)"
FAIL e.style['letter-spacing'] = "calc(clamp(-30px, -100px, -10px))" should set the property value assert_equals: serialization should be canonical expected "calc(-30px)" but got "clamp(-30px, -100px, -10px)"
FAIL e.style['letter-spacing'] = "clamp(-30px, 100px, -10px)" should set the property value assert_equals: serialization should be canonical expected "calc(-10px)" but got "clamp(-30px, 100px, -10px)"
FAIL e.style['letter-spacing'] = "calc(clamp(-30px, 100px, -10px))" should set the property value assert_equals: serialization should be canonical expected "calc(-10px)" but got "clamp(-30px, 100px, -10px)"
FAIL e.style['letter-spacing'] = "clamp(-10px, 100px, -30px)" should set the property value assert_equals: serialization should be canonical expected "calc(-10px)" but got "clamp(-10px, 100px, -30px)"
FAIL e.style['letter-spacing'] = "calc(clamp(-10px, 100px, -30px))" should set the property value assert_equals: serialization should be canonical expected "calc(-10px)" but got "clamp(-10px, 100px, -30px)"
FAIL e.style['letter-spacing'] = "clamp(-10px, -100px, -30px)" should set the property value assert_equals: serialization should be canonical expected "calc(-10px)" but got "clamp(-10px, -100px, -30px)"
FAIL e.style['letter-spacing'] = "calc(clamp(-10px, -100px, -30px))" should set the property value assert_equals: serialization should be canonical expected "calc(-10px)" but got "clamp(-10px, -100px, -30px)"
FAIL e.style['letter-spacing'] = "calc(0px + clamp(10px, 20px, 30px))" should set the property value assert_equals: serialization should be canonical expected "calc(20px)" but got "calc(0px + clamp(10px, 20px, 30px))"
FAIL e.style['letter-spacing'] = "calc(calc(0px + clamp(10px, 20px, 30px)))" should set the property value assert_equals: serialization should be canonical expected "calc(20px)" but got "calc(0px + clamp(10px, 20px, 30px))"
FAIL e.style['letter-spacing'] = "calc(0px - clamp(10px, 20px, 30px))" should set the property value assert_equals: serialization should be canonical expected "calc(-20px)" but got "calc(0px - clamp(10px, 20px, 30px))"
FAIL e.style['letter-spacing'] = "calc(calc(0px - clamp(10px, 20px, 30px)))" should set the property value assert_equals: serialization should be canonical expected "calc(-20px)" but got "calc(0px - clamp(10px, 20px, 30px))"
FAIL e.style['letter-spacing'] = "calc(0px + clamp(30px, 100px, 20px))" should set the property value assert_equals: serialization should be canonical expected "calc(30px)" but got "calc(0px + clamp(30px, 100px, 20px))"
FAIL e.style['letter-spacing'] = "calc(calc(0px + clamp(30px, 100px, 20px)))" should set the property value assert_equals: serialization should be canonical expected "calc(30px)" but got "calc(0px + clamp(30px, 100px, 20px))"
FAIL e.style['letter-spacing'] = "calc(0px - clamp(30px, 100px, 20px))" should set the property value assert_equals: serialization should be canonical expected "calc(-30px)" but got "calc(0px - clamp(30px, 100px, 20px))"
FAIL e.style['letter-spacing'] = "calc(calc(0px - clamp(30px, 100px, 20px)))" should set the property value assert_equals: serialization should be canonical expected "calc(-30px)" but got "calc(0px - clamp(30px, 100px, 20px))"
PASS e.style['letter-spacing'] = "clamp(1px, 1em, 1vh)" should set the property value
PASS e.style['letter-spacing'] = "calc(clamp(1px, 1em, 1vh))" should set the property value
PASS e.style['letter-spacing'] = "calc(0px + clamp(1px, 1em, 1vh))" should set the property value
PASS e.style['letter-spacing'] = "calc(calc(0px + clamp(1px, 1em, 1vh)))" should set the property value
PASS e.style['letter-spacing'] = "calc(0px - clamp(1px, 1em, 1vh))" should set the property value
PASS e.style['letter-spacing'] = "calc(calc(0px - clamp(1px, 1em, 1vh)))" should set the property value
Harness: the test ran to completion.
